Moved functions around to fix a packaging error when executing amd from the package dir; added support for * globbing in a flat dir
